Improved main page, group cards and list functionality
Implemented new menu for main page to make true to Michelle's design and also updated group cards to align with such a design. The Item search view was refined with removed navbar and item add view was improved with new buttons. Finally, settings received a reskin and some new functionality